# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 校园失物招领系统 API
  version: 0.2.0
  description: 终审整改契约：与Controller method+path及实际HTTP行为对应；所有JSON成功为200且含ApiEnvelope，文件下载例外。时间UTC带Z；ID为int64数值。动态文件权限、受理授权见conventions。运行期springdoc用于交叉检查，不能替代此处业务契约。
servers:
- url: /api/v1
  description: 统一前缀
...
security:
- bearerAuth: []
paths:
...
  /admin/disputes:
    get:
      tags:
      - M2-Admin
      summary: 争议列表(FR-DISPUTE-02)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/AdminDisputeView'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: status
        in: query
        schema:
          type: string
  /admin/disputes/{disputeId}:
    get:
      tags:
      - M2-Admin
      summary: 受限查看争议（FR-DISPUTE-02）
      parameters:
      - name: disputeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/AdminDisputeView'
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      description: 未受理管理员只得到列表级字段；description及证据仅当前受理管理员可见。
  /admin/disputes/{disputeId}/assign:
    post:
      tags:
      - M2-Admin
      summary: 受理争议（受理后方可读取受限证据，FR-DISPUTE-02）
      parameters:
      - name: disputeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '409':
          description: DISPUTE_NOT_OPEN
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
  /admin/disputes/{disputeId}/resolution:
    post:
      tags:
      - M2-Admin
      summary: 裁决争议（FR-DISPUTE-02）
      parameters:
      - name: disputeId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveRequest'
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      nullable: true
                      description: 该操作data为null
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      description: 后端强制校验当前管理员为受理人，否则404。
  /admin/audit-logs:
    get:
      tags:
      - M2-Admin
      summary: 审计日志(FR-AUDIT-01)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/AuditLog'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
      - name: targetId
        in: query
        schema:
          type: integer
          format: int64
      - name: targetType
        in: query
        schema:
          type: string
      - name: action
        in: query
        schema:
          type: string
  /admin/maintenance/backups:
    get:
      tags:
      - M2-Admin
      summary: 备份记录(FR-BACKUP-01)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      type: object
                      required:
                      - items
                      - total
                      - page
                      - pageSize
                      properties:
                        items:
                          type: array
                          items:
                            $ref: '#/components/schemas/BackupRecord'
                        total:
                          type: integer
                          format: int64
                        page:
                          type: integer
                          format: int32
                        pageSize:
                          type: integer
                          format: int32
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          maximum: 100000
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
          maximum: 100
    post:
      tags:
      - M2-Admin
      summary: 触发授权备份(FR-BACKUP-01)
      responses:
        '200':
          description: OK；统一响应封装
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/ApiEnvelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/BackupRecord'
        '400':
          description: 参数格式或业务校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          description: 未登录或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '403':
          description: 角色或账户受限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '404':
          description: 不存在或无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '409':
          description: 状态/唯一约束冲突
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiEnvelope:
      type: object
      properties:
        code:
          type: string
          example: OK
        message:
          type: string
        data: {}
        requestId:
          type: integer
          nullable: true
          format: int64
...
    ResolveRequest:
      type: object
      properties:
        resolutionType:
          type: string
          enum:
          - CONTINUE
          - TERMINATE_REOPEN
          - CLOSE
        resolutionNote:
          type: string
          maxLength: 500
          minLength: 1
      required:
      - resolutionType
      - resolutionNote
...
    AdminDisputeView:
      type: object
      properties:
        id:
          type: integer
          format: int64
          nullable: true
        claimId:
          type: integer
          format: int64
          nullable: true
        raisedBy:
          type: integer
          format: int64
          nullable: true
        reason:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        assignedAdminId:
          type: integer
          format: int64
          nullable: true
        resolutionType:
          type: string
          nullable: true
        resolutionNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        resolvedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
          nullable: true
        evidenceFileIds:
          type: array
          items:
            type: integer
            format: int64
...
    BackupRecord:
      type: object
      properties:
        id:
          type: integer
          format: int64
        initiatedBy:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        finishedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        status:
          type: string
        manifestPath:
          type: string
        checksum:
          type: string
        restoreVerifiedAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
...
    AuditLog:
      type: object
      properties:
        id:
          type: integer
          format: int64
        actorId:
          type: integer
          format: int64
        actorType:
          type: string
        action:
          type: string
        targetType:
          type: string
        targetId:
          type: integer
          format: int64
        requestId:
          type: integer
          format: int64
        result:
          type: string
        metadata:
          type: string
        createdAt:
          type: string
          format: date-time
          description: UTC ISO 8601，响应带Z
...